requirements.txt
Django
git+https://github.com/TheEnbyperor/django-keycloak-auth@a592e039784be9c5b5760adeedb66196e8e77591
sentry-sdk
psycopg2-binary
gunicorn
django-crispy-forms
django-xff
djangorestframework
django-cockroachdb==3.1.*
